+++ apps/editors/decoration/constant.js
@@ -1,0 +1,6 @@
+export const MODE = {
+  VIEW: "VIEW",
+  ADD_2D: "ADD_2D",
+  ADD_3D: "ADD_3D",
+  TRANSFORM: "TRANSFORM",
+};