+++ components/Icon/index.js
@@ -7,8 +7,8 @@ const avoidParentHeightPadding = `
 
 export const Image = styled.div`
   ${avoidParentHeightPadding}
-  width: 24px;
-  height: 24px;
+  width: 100%;
+  height: 100%;
   display: inline-block;
   cursor: pointer;
   ${(props) =>
@@ -19,8 +19,20 @@ export const Image = styled.div`
     `};
 `;
 
+export const Wrapper = styled.div`
+  width: 24px;
+  height: 24px;
+  background: ${(props) => (props.$highlight ? "white" : "transparent")};
+  border-radius: 50%;
+  padding: 5px;
+`;
+
 const Icon = ({ src, ...props }) => {
-  return <Image src={src} {...props} alt={src} />;
+  return (
+    <Wrapper {...props}>
+      <Image src={src} {...props} alt={src} />
+    </Wrapper>
+  );
 };
 
 // all svg resources is download from https://www.svgrepo.com/vectors/cursor/
@@ -33,6 +45,10 @@ const files = {
   activated3D: `${IconFolder}/activated3D.svg`,
   inactivated3D: `${IconFolder}/inactivated3D.svg`,
   share: `${IconFolder}/share.svg`,
+  cursor: `${IconFolder}/cursor.svg`,
+  placeholder: `${IconFolder}/placeholder.svg`,
+  box: `${IconFolder}/box.svg`,
+  axis: `${IconFolder}/axis.svg`,
 };
 
 const Icons = Object.keys(files).reduce((acc, key) => {