+++ three/constant/index.js
@@ -1,0 +1,4 @@
+export const RENDER_ORDER = {
+  CSS3D: 0,
+  MESH: 1,
+};