+++ three/index.js
@@ -4,11 +4,18 @@ import ProjectionMesh from "./components/PanoramaProjectionMesh";
 import TextureMesh from "./components/PanoramaTextureMesh";
 import Canvas from "./components/ThreeCanvas";
 import Math from "./core/helpers/Math";
-import { raycastGeometry } from "./core/helpers/Raycaster";
+import {
+  raycastGeometry,
+  raycastMeshFromScreen,
+} from "./core/helpers/Raycaster";
 import { getEmptyRoomGeometry, downloadMesh } from "./core/RoomGeometry";
+import THREECss3DObject from "./components/Css3DObject";
+import THREEPlaceholder from "./components/Placeholder";
 
 // React components
 export const ThreeCanvas = Canvas;
+export const Css3DObject = THREECss3DObject;
+export const Placeholder = THREEPlaceholder;
 export const PanoramaOutline = THREEPanoramaOutline;
 export const PanoramaProjectionMesh = ProjectionMesh;
 export const PanoramaTextureMesh = TextureMesh;
@@ -20,6 +27,7 @@ export const Loaders = hooks;
 export const Core = {
   Math,
   raycastGeometry,
+  raycastMeshFromScreen,
   getEmptyRoomGeometry,
   downloadMesh,
 };